+++ guillotina/__init__.py
@@ -13,6 +13,7 @@ _ = MessageFactory('guillotina')
 
 
 app_settings = {
+    "debug": False,
     "aiohttp_settings": {},
     "databases": [],
     "conflict_retry_attempts": 3,